\name{tswv}
\docType{data}
\alias{tswv}
\title{Tomato Spotted Wilt Virus data}
\description{
  Data extracted from plots in Hughes et al. (1997)
}
\usage{data(tswv)}
\format{
  A list of two objects of class ``datagen'' as follows:
   \describe{
    \item{tswvsir:}{an object of class ``datagen'' for the ``SIR'' compartmental framework.}
    \item{tswvsinr:}{an object of class ``datagen'' for the ``SINR'' compartmental framework.}
  }
}
\references{  
  Hughes, G., McRoberts, N., Madden, L.V., Nelson, S. C. (1997). Validating mathematical models of plant disease progress in space and time. \emph{IMA Journal of Mathematics Applied in Medicine and Biology}, 14, 85-112.
}
\keyword{datasets}