\name{tswv}
\docType{data}
\alias{tswv}
\title{Tomato Spotted Wilt Virus data}
\description{
  Data extracted from plots in Hughes et al. (1997)
}
\usage{tswv}
\format{
  A list of data frames as follows:
   \describe{
    \item{loc}{The XY coordinates of plants.}
    \item{tswvsir}{The tswv data in the format of 4 columns: id numbers of plants, removal times, infectious periods, and infection times.}
    \item{tswvsinr}{The tswv data in the format of 6 columns: id numbers of plants, removal times, delay periods, notification times, incubation periods, and infection times.}
  }
}
\references{
  
  Hughes, G., McRoberts, N., Madden, L.V., Nelson, S. C. (1997). Validating mathematical models of plant disease progress in space and time. \emph{IMA Journal of Mathematics Applied in Medicine and Biology}, 14, 85-112.
}
\keyword{datasets}